Fix: Use collection_name for the trackable scope.
When sub-classing a trackable model, using the class name for the scope
was resulting in errors. The problem seems to be that the name of the
class was different in different contexts. (e.g., inclusion in the
parent class vs execution in the sub-class)

Switching to using the collection_name should make the scope consistent
across all calls.