Add helper to track all embedded relations.
I think this was already in the works given the deprecation warning
when 'on' is set to 'all'.  Adding 'fields' already triggers on all
of the fields of the model, now adding 'relations' will do the same
for all embedded relations.  ['fields', 'relations'] should perform
the desired future behavior of 'all'.