No-op on repeated calls to destroy.
Calling destroy multiple times on the same document attempts to track each deletion, which ultimately fails with a RuntimeError when updating the (frozen) attributes hash with the incremented version number. (This is especially problematic / difficult to track down for documents with cyclic relations and cascading destroys).

This no-ops tracking of destroy for documents that are already destroyed.