Avoid adding the version attribute to deleted docs
I had to relax a couple of the rubocop constraints in order to fit these
changes in.

Also, there was another change in integration_spec.rb that appears
unrelated. After making the change in trackable.rb, the specs started
failing because, while the #history_tracks method was returning a new
array with an additional member, the #last method was continuing to
return the same record. Addressing the last member of the array by index
avoids the issue.