Test setting track_blank_changes option
There are a number of tests that setting an option value actually sets the
option to that value, so this commit adds one for track_blank_changes.
Unfortunately I added it before discovering that the whole section isn't
testing anything other than that setting a value in a Hash and retrieving
that value gives the same value back (and that two of tests even do that
wrong).

The two tests that are wrong are those for :track_create and :track_destroy.
Because they're setting the options to the default value, it's not possible
to tell if the tests pass because the values are actually being set or
because they're not and just the default value is being used.

More problematic is that adding a test like:

        describe ':any_name' do
          let(:options) { { any_name: 'any_value' } }
          it { expect(subject[:any_name]).to be 'any_value' }
        end

passes, as long as any_name is something other than a few special option
names (like on). None of the tests between lines 319 and 359 are really
testing anything useful IMO.

Nevertheless, I added the :track_blank_changes test. Here's the diff of the
test output vs that generated after commit 8081c34:

  $ diff -u testresults/2-additional-all-blank-changes.txt testresults/3-set-track_blank_changes.txt
  --- testresults/2-additional-all-blank-changes.txt	2024-08-05 11:34:14.000000000 -0700
  +++ testresults/3-set-track_blank_changes.txt	2024-08-05 15:46:06.000000000 -0700
  @@ -591,6 +591,8 @@
	     is expected to equal false
	   :track_destroy
	     is expected to equal true
  +        :track_blank_changes
  +          is expected to equal true
	   #remove_reserved_fields
	     is expected to eq ["foo"]
	     is expected to eq []
  @@ -1076,7 +1078,7 @@
	 # /Users/blm/.rvm/gems/ruby-2.3.7/gems/rspec-core-3.13.0/exe/rspec:4:in `<main>'

   Finished in n minute n seconds (files took n seconds to load)
  -455 examples, 14 failures, 2 pending
  +456 examples, 14 failures, 2 pending

   Failed examples: